requirements.txt
# Core ML dependencies
ultralytics

# API server
fastapi
uvicorn
python-multipart

# Image processing
Pillow
opencv-python

# Plotting
matplotlib
seaborn
